Extract CaaS payload builder into leaf module + standalone bundle
- Add buildCaasXdmPayload() + hasCardMetadata/isDisabledOnPage in a new
  self-contained leaf (tools/send-to-caas/caas-payload-core.js) that does
  not import utils.js/caas-utils.js, so it can be bundled standalone.
- send-utils.js re-exports the payload surface and keeps the browser-only
  postDataToCaaS/isPagePublished; injects the lazy 706KB caas-tags fallback
  so the leaf stays free of it.
- Refactor caasAutoPublish() to use buildCaasXdmPayload().
- Add build/caas-payload-builder.js entry + build:caas-payload-builder
  microbundle script producing a single ~20KB UMD bundle for milo-caas to
  fetch and run.

271 CaaS unit tests pass.
