+++ package.json
@@ -18,6 +18,7 @@
     "build:htm-preact": "microbundle ./build/htm-preact.js -o ./libs/deps/htm-preact.js -f modern --no-sourcemap --target web; mv ./libs/deps/htm-preact.modern.mjs ./libs/deps/htm-preact.js",
     "build:preact-debug": "microbundle ./build/htm-preact-debug.js -o ./libs/deps/htm-preact-debug.js -f modern --no-sourcemap --target web; mv ./libs/deps/htm-preact-debug.modern.mjs ./libs/deps/htm-preact-debug.js",
     "build:gnav-profile": "microbundle ./libs/blocks/global-navigation/blocks/profile/profile-wrapper.js -o ./build/profile-wrapper-build.js -f umd --external none --no-sourcemap --target web; mv ./build/profile-wrapper-build.umd.js ./build/profile.js",
+    "build:caas-payload-builder": "microbundle ./build/caas-payload-builder.js -o ./tools/send-to-caas/caas-payload-builder.js -f umd --external none --no-sourcemap --target web; mv ./tools/send-to-caas/caas-payload-builder.umd.js ./tools/send-to-caas/caas-payload-builder.js",
     "build:echarts": "cp node_modules/echarts/dist/echarts.common.min.js libs/deps/echarts.common.min.js",
     "build:s2a": "node .github/workflows/update-s2a-styles.js"
   },
