+++ tools/send-to-caas/caas-payload-builder.js
@@ -1,0 +1,1 @@
+!function(t,e){"object"==typeof exports&&"undefined"!=typeof module?e(exports):"function"==typeof define&&define.amd?define(["exports"],e):e((t||self).milo={})}(this,function(t){const e="0123456789abcdef".split(""),n=t=>{const n=t>>4;return e[n]+e[t-(n<<4)]},r=t=>[...t].map(t=>n(t)).join(""),i=function(t){return Promise.resolve(function(t){try{const e=(new TextEncoder).encode(t);return Promise.resolve(crypto.subtle.digest("SHA-1",e))}catch(t){return Promise.reject(t)}}(t)).then(function(t){return(t=>[r(t.slice(0,4)),"-",r(t.slice(4,6)),"-",n(15&t[6]|parseInt(50,16)),n(t[7]),"-",n(63&t[8]|128),n(t[9]),"-",r(t.slice(10,16))].join(""))(new Uint8Array(t))})};function o(t,e){try{var n=t()}catch(t){return e(t)}return n&&n.then?n.then(void 0,e):n}function a(t,e,n){if(!t.s){if(n instanceof c){if(!n.s)return void(n.o=a.bind(null,t,e));1&e&&(e=n.s),n=n.v}if(n&&n.then)return void n.then(a.bind(null,t,e),a.bind(null,t,2));t.s=e,t.v=n;const r=t.o;r&&r(t)}}const c=/*#__PURE__*/function(){function t(){}return t.prototype.then=function(e,n){const r=new t,i=this.s;if(i){const t=1&i?e:n;if(t){try{a(r,1,t(this.v))}catch(t){a(r,2,t)}return r}return this}return this.o=function(t){try{const i=t.v;1&t.s?a(r,1,e?e(i):i):n?a(r,1,n(i)):a(r,2,i)}catch(t){a(r,2,t)}},r},t}();function s(t){return t instanceof c&&1&t.s}const l="undefined"!=typeof Symbol?Symbol.iterator||(Symbol.iterator=Symbol("Symbol.iterator")):"@@iterator",u=/(http(s)?:\/\/.)?(www\.)?[-a-zA-Z0-9@:%._+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_+.~#?&//=]*)/,f=/fragments(.*)#[a-zA-Z0-9_-]+$/,d={cc:"hawks",dc:"doccloud"},h=/(\s*\S+\s*:\s*\S+\s*)/,m=t=>f.test(t),p="langstore",g="target-preview",v=new URL(window.location.href);function y(t,e){void 0===e&&(e=document);const n=t&&t.includes(":")?"property":"name",r=e.head.querySelector(`meta[${n}="${t}"]`);return r&&r.content}function b(t,e){if(void 0===e&&(e=window.location.pathname),!t)return{ietf:"en-US",tk:"hah7vzn.css",prefix:""};const n=e.split("/"),r=n[1],i=[p,g].includes(r)?r:"",o=n[2];let a="";i?a=Object.keys(t).find(e=>t[e]?.ietf?.startsWith(o))??"":r in t&&(a=r);const c=function(t,e){const n=t[e],r=(t,e)=>({...t,prefix:e?`/${e}`:"",region:t.region||e.split("_")[0]||"us"});if(!("base"in n)){const i=Object.entries(t).filter(t=>{let[,n]=t;return n.base===e}).reduce((t,e)=>{let[i,o]=e;const a={...n,...o,base:o.base};return t[i]=r(a,i),t},{});return{...r(n,e),regions:i}}if("base"in n&&void 0!==t[n.base]){const i={...t[n.base],...n};return r(i,e)}return{...n}}(t,a);return i&&(c.prefix=`/${i}${o?`/${o}`:""}`),c}const w={en:"en",de:"de",fr:"fr","fr-ca":"fr-ca",ja:"ja",ar:"ar",arabic:"ar",bg:"bg",cs:"cs",da:"da",es:"es",et:"et",fi:"fi",he:"he",hu:"hu",it:"it",ko:"ko",lt:"lt",lv:"lv",nl:"nl",no:"no",pl:"pl",pt:"pt",ro:"ro",ru:"ru",sk:"sk",sl:"sl",sv:"sv",tr:"tr",uk:"uk","zh-hant":"zh-hant",th:"th",fil:"fil",id:"id",ms:"ms",vi:"vi",hi:"hi",el:"el","":"en"},x={ar:{ietf:"es-AR"},br:{ietf:"pt-BR"},ca:{ietf:"en-CA"},ca_fr:{ietf:"fr-CA"},cl:{ietf:"es-CL"},co:{ietf:"es-CO"},cr:{ietf:"es-CR"},ec:{ietf:"es-EC"},el:{ietf:"es-EL"},gt:{ietf:"es-GT"},la:{ietf:"es-LA"},mx:{ietf:"es-MX"},pe:{ietf:"es-PE"},pr:{ietf:"es-PR"},"":{ietf:"en-US"},langstore:{ietf:"en-US"},africa:{ietf:"en-africa"},gb:{ietf:"xx-gb"},be_fr:{ietf:"fr-BE"},be_en:{ietf:"en-BE"},be_nl:{ietf:"nl-BE"},be:{ietf:"xx-be"},cy_en:{ietf:"en-CY"},cy:{ietf:"xx-cy"},dk:{ietf:"da-DK"},de:{ietf:"de-DE"},ee:{ietf:"et-EE"},eg_ar:{ietf:"ar-EG"},eg_en:{ietf:"en-GB"},eg:{ietf:"xx-eg"},es:{ietf:"es-ES"},fr:{ietf:"fr-FR"},gr_en:{ietf:"en-GR"},gr_el:{ietf:"el-GR"},gr:{ietf:"xx-gr"},ie:{ietf:"en-IE"},il_en:{ietf:"en-IL"},il_he:{ietf:"he-il"},il:{ietf:"xx-il"},it:{ietf:"it-IT"},kw_ar:{ietf:"ar-KW"},kw_en:{ietf:"en-GB"},kw:{ietf:"xx-kw"},lv:{ietf:"lv-LV"},lt:{ietf:"lt-LT"},lu_de:{ietf:"de-LU"},lu_en:{ietf:"en-LU"},lu_fr:{ietf:"fr-LU"},lu:{ietf:"xx-lu"},hu:{ietf:"hu-HU"},mt:{ietf:"en-MT"},mena_en:{ietf:"en-mena"},mena_ar:{ietf:"ar-mena"},mena:{ietf:"xx-mena"},ng:{ietf:"en-NG"},nl:{ietf:"nl-NL"},no:{ietf:"no-NO"},pl:{ietf:"pl-PL"},pt:{ietf:"pt-PT"},qa_ar:{ietf:"ar-QA"},qa_en:{ietf:"en-GB"},qa:{ietf:"xx-qa"},ro:{ietf:"ro-RO"},sa_en:{ietf:"en-sa"},ch_fr:{ietf:"fr-CH"},ch_de:{ietf:"de-CH"},ch_it:{ietf:"it-CH"},ch:{ietf:"xx-ch"},si:{ietf:"sl-SI"},sk:{ietf:"sk-SK"},fi:{ietf:"fi-FI"},se:{ietf:"sv-SE"},tr:{ietf:"tr-TR"},ae_en:{ietf:"en-ae"},uk:{ietf:"en-GB"},at:{ietf:"de-AT"},cz:{ietf:"cs-CZ"},bg:{ietf:"bg-BG"},ru:{ietf:"ru-RU"},cis:{ietf:"xx-cis"},ua:{ietf:"uk-UA"},ae_ar:{ietf:"ar-ae"},ae:{ietf:"xx-ae"},sa_ar:{ietf:"ar-sa"},sa:{ietf:"xx-sa"},za:{ietf:"en-ZA"},apac:{ietf:"xx-apac"},hk:{ietf:"xx-hk"},au:{ietf:"en-AU"},hk_en:{ietf:"en-HK"},in:{ietf:"en-in"},id_id:{ietf:"id-id"},id_en:{ietf:"en-id"},id:{ietf:"xx-id"},my_ms:{ietf:"ms-my"},my_en:{ietf:"en-my"},my:{ietf:"xx-my"},nz:{ietf:"en-nz"},ph_en:{ietf:"en-ph"},ph_fil:{ietf:"fil-PH"},ph:{ietf:"xx-ph"},sg:{ietf:"en-SG"},th_en:{ietf:"en-th"},in_hi:{ietf:"hi-in"},th_th:{ietf:"th-th"},th:{ietf:"xx-th"},cn:{ietf:"zh-CN"},hk_zh:{ietf:"zh-HK"},tw:{ietf:"zh-TW"},jp:{ietf:"ja-JP"},kr:{ietf:"ko-KR"},vn_en:{ietf:"en-vn"},vn_vi:{ietf:"vi-VN"},vn:{ietf:"xx-vn"},sea:{ietf:"xx-sea"}},k=[];function C(t,e){void 0===e&&(e=k);const n=t.split("/")[1];return e.includes(n)?n:""}const P=function(t,e){void 0===t&&(t=window.location?.hostname||""),void 0===e&&(e=window.location?.pathname||"");const n=/^[^.]+\.([a-z-]+-)?graybox\.adobe\.com$/.test(t),r=(t.endsWith(".aem.page")||t.endsWith(".aem.live"))&&t.includes("graybox");if(n){const e=t.split(".");if(e.length>=3&&("graybox"===e[1]||e[1].includes("-graybox")))return e[0]}if(r){const t=e.split("/").filter(Boolean);if(t.length>0)return t[0]}return null},$=(t,e,n)=>{if(!t)return!1;const r=t.split(",").map(t=>t.trim().replace(/^\//,""));return r.includes(e)||Boolean(n)&&r.includes(`${e}_${n}`)};let S;const _=function(t,e,n){try{function r(){return{country:c.toLowerCase(),lang:a.toLowerCase(),...s&&{fromFallback:!0}}}const i=t.split("/");let a="en",c="xx",s=!1;const l=function(){if("news"!==e.toLowerCase())return Promise.resolve(function(t,e,n){void 0===n&&(n="www.adobe.com");try{let r;const i=t.toLowerCase();let a={country:"xx",language:"en"},c=e;if(e.includes("://")||!e.startsWith("/"))try{c=new URL(e.startsWith("http")?e:`https://${e}`).pathname}catch(t){e.startsWith("/")||(c=`/${e.split("/").slice(1).join("/")}`)}const s=c.split("/")[1],l=o(function(){let t;return Promise.resolve(function(t){return void 0===t&&(t="www.adobe.com"),S||(S=fetch(`https://www.adobe.com/federal/assets/data/lingo-site-mapping.json?${t}`).then(t=>{if(!t.ok)throw new Error(`HTTP ${t.status}`);return t.json()})),S}(n)).then(function(e){const n=e["site-query-index-map"]?.data??[],o=e["site-locales"]?.data??[],c=n.find(t=>{let{caasOrigin:e}=t;return i===e});if(c&&(t=c.uniqueSiteId),!o.some(e=>{let{uniqueSiteId:n,baseSite:r,regionalSites:i}=e;return n===t&&(s===r.split("/")[1]||$(i,s))})){const t=x[s]?.ietf||"en-US";let[e,n]=t.split("-");return r=1,{country:n,language:e}}return o.filter(e=>{let{uniqueSiteId:n}=e;return n===t}).forEach(t=>{let{baseSite:e,regionalSites:n}=t;s!==e.split("/")[1]?$(n,s)&&("/"===e&&(a={country:s,language:"en"}),a={country:s,language:e.split("/")[1]}):a={country:"xx",language:e.split("/")[1]}}),r=1,a})},function(t){window.lana?.log(`Failed to load lingo-site-mapping.json: ${t}`,{tags:"caas",severity:"error"}),a.fromFallback=!0});return Promise.resolve(l&&l.then?l.then(function(t){return r?t:a}):r?l:a)}catch(t){return Promise.reject(t)}}(e,t,n)).then(function(t){s=!0===t.fromFallback,c=x[t.country.toLowerCase()]??"xx","object"==typeof c&&(c=c.ietf?.split("-")[1]??"xx"),a=t.language??"en"});a=w[i[1]]??w[""]??"en",c=x[i[2]]??"xx","object"==typeof c&&(c=c.ietf?.split("-")[1]??"xx")}();return Promise.resolve(l&&l.then?l.then(r):r())}catch(u){return Promise.reject(u)}},[L,j]=(()=>{let t={isInjectedDoc:()=>(void 0).doc!==document,doc:document};return[e=>(t={...t,...e},t),()=>t]})(),U=t=>{if(t.startsWith("http"))return t;const{host:e}=j();return`https://${e}${t.startsWith("/")?"":"/"}${t}`},E=(t,e)=>{const n=j().doc.querySelector(`meta[${t}='${e}']`);if(n)return n.content},I=(t,e)=>{if(void 0===t||m(t))return t;const n=t.includes("href=")?(t=>{const e=document.createElement("div");return e.innerHTML=t,e.querySelector("a").getAttribute("href")})(t):t;return m(n)?n:u.test(n)?(t=>t?.startsWith("https://")||t?.startsWith("http://")?t:`https://${t}`)(n):{error:e}},q=function(t,e,n){void 0===n&&(n=[]);const r=[];let i=Object.values(t).find(t=>!(n.includes(t.title)||n.includes(t.name)||n.includes(t.path)||n.includes(t.tagID)||(t.tags&&Object.keys(t.tags).length&&r.push(t.tags),![t.title.toLowerCase(),t.name,t.path,t.path.replace("/content/cq:tags/",""),t.tagID.toLowerCase()].includes(e.toLowerCase()))));return i||r.some(t=>(i=q(t,e,n),i)),i},[D,T]=(()=>{let t;return[()=>t,function(){try{let n;function e(e){if(n)return e;const{getCaasTagsFallback:r}=j(),i=function(){if(r)return Promise.resolve(r()).then(function(e){t=e?.namespaces?.caas?.tags})}();return i&&i.then?i.then(function(){}):void 0}const r=o(function(){return Promise.resolve(fetch("https://www.adobe.com/chimera-api/tags")).then(function(e){return function(){if(e.ok)return Promise.resolve(e.json()).then(function(e){t=e.namespaces.caas.tags,n=1})}()})},function(){});return Promise.resolve(r&&r.then?r.then(e):e(r))}catch(i){return Promise.reject(i)}}]})(),R=(t,e)=>{if(!t)return;const n=D(),r=q(n,t,["caas:events"])||q(n.events.tags,t,[]);return r||e.push(t),r},M=(t,e)=>{if(t)try{const n=new Date(t);return n.getFullYear()<2e3?{error:`${e} - Date is before the year 2000`}:n.toISOString()}catch(t){return{error:e}}},A=(t,e)=>{if(t&&e&&"default"!==e){const n=`-fg-${e}`;if(t.endsWith(n))return t.slice(0,t.lastIndexOf(n))}return t},W=t=>{const{project:e,repo:n}=j(),r=e||A(n,t),i=d[r.toLowerCase()]||r;if(i)return i;if(window.location.hostname.endsWith(".page")){const[,e]=window.location.hostname.split(".")[0].split("--");return A(e,t)}throw new Error("No Project or Repo defined in config")},z=t=>{const e=j().doc.querySelector(".card-metadata");if(!e)return null;let n="";return[...e.children].some(e=>{const r=e.firstElementChild.textContent?.trim().toLowerCase();if(r!==t)return!1;const i=e.lastElementChild.querySelector("img");if(i){let t=i.src;if(j().bulkPublish){const e=i.attributes.src.value;if(e.startsWith("./")){const n=(t=>`${t.split("/").slice(0,-1).join("/")}/`)(j().pageUrl);t=`${n}${e}`}else t=e.startsWith("/")?`${new URL(j.pageUrl.origin)}${e}`:e}n=new URL(t)?.pathname}else n=e.lastElementChild.textContent?.trim();return!0}),n},O=()=>{const{doc:t}=j(),e=z("image")||z("cardimage")||z("cardimagepath")||t.querySelector("main")?.querySelector("img")?.src.replace(/\?.*/,"")||t.querySelector('meta[property="og:image"]')?.content;return e?U(e):null},F=t=>{const e=[];return t.badgeimage&&e.push({type:"image",value:U(t.badgeimage)}),t.badgetext&&e.push({type:"text",value:t.badgetext}),e},G=1e3,B=function(t,e){try{function n(t){const e=t?.toLowerCase().split("-")||[],[n="en",r="us"]=e;return{country:r,lang:n}}if(function(){const t=(v.searchParams.get("langfirst")||y("langfirst"))?.toLowerCase();return["true","on"].includes(t)}()){const i=window.location.pathname.includes("/tools/send-to-caas/bulkpublisher")?"bulkpublisher":window.location.hostname;return _(window.location.pathname,d[e.toLowerCase()]||e,i)}const r=window.location.pathname.includes("/tools/send-to-caas/bulkpublisher");return Promise.resolve(r?Promise.resolve(function(t){try{let{getLocale:e}=j();return t.languageFirst?function(t,e,n){void 0===n&&(n={});try{let e;const{retryDelayMs:r=G}=n;let i,o=0;const l=function(t,e,n){for(var r;;){var i=t();if(s(i)&&(i=i.v),!i)return o;if(i.then){r=0;break}var o=n();if(o&&o.then){if(!s(o)){r=1;break}o=o.s}if(e){var l=e();if(l&&l.then&&!s(l)){r=2;break}}}var u=new c,f=a.bind(null,u,2);return(0===r?i.then(h):1===r?o.then(d):l.then(m)).then(void 0,f),u;function d(r){o=r;do{if(e&&(l=e())&&l.then&&!s(l))return void l.then(m).then(void 0,f);if(!(i=t())||s(i)&&!i.v)return void a(u,1,o);if(i.then)return void i.then(h).then(void 0,f);s(o=n())&&(o=o.v)}while(!o||!o.then);o.then(d).then(void 0,f)}function h(t){t?(o=n())&&o.then?o.then(d).then(void 0,f):d(o):a(u,1,o)}function m(){(i=t())?i.then?i.then(h).then(void 0,f):h(i):a(u,1,o)}}(function(){return!e&&o<=3},function(){return!!(o+=1)},function(){function n(){return Promise.resolve((n=t.prodUrl,r=t.repo,o=t.host,_(n,d[r.toLowerCase()]||r,o))).then(function(t){const{country:n,lang:r,fromFallback:o}=t;if(i=`${r}-${n}`,"en"!==r||"xx"!==n||!o)return e=1,i});var n,r,o}const a=function(){if(o>0)return Promise.resolve((t=r,new Promise(e=>{setTimeout(e,t)}))).then(function(){});var t}();return a&&a.then?a.then(n):n()});return Promise.resolve(l&&l.then?l.then(function(t){return e?t:i}):e?l:i)}catch(t){return Promise.reject(t)}}(t):(e||(e=b,L({getLocale:e})),Promise.resolve(e(x,t.prodUrl).ietf))}catch(t){return Promise.reject(t)}}(t)).then(n):n((x[window.location.pathname.split("/")[1]]||x[""]).ietf))}catch(o){return Promise.reject(o)}},H=()=>{const t={},e=j().doc.querySelector(".caas-marquee-metadata"),n=j().doc.querySelector(".card-metadata")||e,r=["description"];return n&&n.childNodes.forEach(n=>{const i=n.children?.[0]?.textContent?.toLowerCase();let o=n.children?.[1]?.textContent;e&&r.includes(i)&&(o=n.children?.[1]?.innerHTML),i&&(t[i]=o)}),t};function N(t,e,n){if((""===t?.trim()||void 0===t)&&n>1)return"";const r=""!==t?.trim()&&void 0!==t?t:e.prodUrl||window.location.origin+window.location.pathname;return I(r,`Invalid Cta${n}Url: ${r}`)}function K(t){if("string"!=typeof t||""===t.trim())return t;try{const e="true"===(y("caaslocaleinject")||"").toLowerCase(),n=C(window.location.pathname);if(!e||!n)return t;const r=new URL(t,window.location.origin);if(!C(r.pathname))return r.pathname=`/${n}${r.pathname}`,r.toString()}catch{}return t}const Z={arbitrary:t=>(t=>t?t.split(",").filter(t=>t.length).filter(t=>h.test(t)).map(t=>{const[e,...n]=t.split(":");return{[e.trim()]:n.join(":").trim()}}):[])(t).map(t=>t),badgeimage:()=>z("badgeimage"),badgetext:0,bookmarkaction:0,bookmarkenabled:function(t){if(void 0===t&&(t=""),t){const e=t.toLowerCase();if("true"===e||"on"===e||"yes"===e)return!0}},bookmarkicon:0,carddescription:0,cardtitle:0,cardimage:()=>O(),cardimagealttext:t=>t||(()=>{if(H().cardimage)return"";const t=O(),e=new URL(t).pathname.split("/").pop(),n=j().doc.querySelector(`img[src*="${e}"]`);return n?.alt})(),contentid:(t,e)=>i(e.prodUrl),contenttype:t=>t||E("property","og:type")||j().contentType,country:function(t,e){try{if(t)return Promise.resolve(t);const n=e.floodgatecolor||y("floodgatecolor"),r=W(n);return Promise.resolve(B(e,r)).then(function(t){let{country:e}=t;return e})}catch(t){return Promise.reject(t)}},created:t=>{if(t)return M(t,`Invalid Created Date: ${t}`);const e=H()?.carddate;if(e)return M(e,`Invalid Date: ${e}`);const n=E("name","publishdate")||E("name","publication-date"),{doc:r,lastModified:i}=j();return n?M(n,`publication-date metadata is not a valid date: ${n}`):M(i||r.lastModified,`document.lastModified is not a valid date: ${r.lastModified}`)},cta1icon:t=>I(t,`Invalid Cta1Icon url: ${t}`),cta1style:0,cta1target:0,cta1text:0,cta1url:(t,e)=>K(N(t,e,1)),cta2icon:t=>I(t,`Invalid Cta2Icon url: ${t}`),cta2style:0,cta2target:0,cta2text:0,cta2url:t=>K(N(t,{},2)),description:t=>t||E("name","description")||"",details:0,entityid:(t,e)=>{const n=e.floodgatecolor||y("floodgatecolor")||"";return i(`${e.prodUrl}${"default"===n||""===n?"":n}`)},env:t=>t||"",eventduration:0,eventend:t=>M(t,`Invalid Event End Date: ${t}`),eventstart:t=>M(t,`Invalid Event Start Date: ${t}`),floodgatecolor:(t,e)=>t||e.floodgatecolor||y("floodgatecolor")||"default",lang:function(t,e){try{if(t)return Promise.resolve(t);const n=e.floodgatecolor||y("floodgatecolor"),r=W(n);return Promise.resolve(B(e,r)).then(function(t){let{lang:e}=t;return e})}catch(t){return Promise.reject(t)}},modified:t=>{const{doc:e,lastModified:n}=j();return t?M(t,`Invalid Modified Date: ${t}`):M(n||e.lastModified,`document.lastModified is not a valid date: ${e.lastModified}`)},origin:(t,e)=>{if(t)return t;const n=e.floodgatecolor||y("floodgatecolor");return W(n)},playurl:t=>I(t,`Invalid PlayURL: ${t}`),primarytag:t=>{const e=R(t);return e?{id:e.tagID}:{}},style:t=>t||"default",tags:t=>(t=>{let e=[];t&&(e=t.toLowerCase().split(/,|(\s+)|(\\n)|;/g).filter(t=>t&&t.trim()&&"\n"!==t));const n=[],r=e.map(t=>R(t,n)).filter(t=>void 0!==t).map(t=>t.tagID),i=[...new Set(r)].map(t=>({id:t}));return{tagErrors:n,tags:i}})(t),title:t=>t||E("property","og:title")||"",uci:(t,e)=>t||e.prodUrl||window.location.pathname,url:(t,e)=>{const n=t||e.prodUrl||window.location.origin+window.location.pathname;return I(n,`Invalid URL: ${n}`)}};t.buildCaasXdmPayload=function(t){let{dom:e,pageUrl:n,lastModified:r,host:i,repo:o,floodgatecolor:u="default",languageFirst:f}=void 0===t?{}:t;try{return L({bulkPublish:!0,doc:e,pageUrl:n,lastModified:r,host:i,repo:o}),Promise.resolve(T()).then(function(){return Promise.resolve(function(t){try{return function(t,e){try{function n(){return!r.contenttype&&u.length&&(r.contenttype=u.find(t=>t.startsWith("caas:content-type"))),{caasMetadata:r,errors:i,tags:u,tagErrors:o}}const r={},i=[];let o=[],u=[];const f=function(t,e){if("function"==typeof t[l]){var n,r,i,o=t[l]();if(function t(l){try{for(;!(n=o.next()).done;)if((l=e(n.value))&&l.then){if(!s(l))return void l.then(t,i||(i=a.bind(null,r=new c,2)));l=l.v}r?a(r,1,l):r=l}catch(t){a(r||(r=new c),2,t)}}(),o.return){var u=function(t){try{n.done||o.return()}catch(t){}return t};if(r&&r.then)return r.then(u,function(t){throw u(t)});u()}return r}if(!("length"in t))throw new TypeError("Object is not iterable");for(var f=[],d=0;d<t.length;d++)f.push(t[d]);return function(t,e){var n,r,i=-1;return function o(l){try{for(;++i<t.length;)if((l=e(i))&&l.then){if(!s(l))return void l.then(o,r||(r=a.bind(null,n=new c,2)));l=l.v}n?a(n,1,l):n=l}catch(t){a(n||(n=new c),2,t)}}(),n}(f,function(t){return e(f[t])})}(Object.entries(Z),function(n){let[a,c]=n;function s(t){t?.error?i.push(t.error):void 0!==t?.tagErrors?(o=t.tagErrors,r[a]=t.tags,u=t.tags.map(t=>t.id)):void 0!==t&&(r[a]=t)}return c?Promise.resolve(c(t[a],e)).then(s):s(t[a])});return Promise.resolve(f&&f.then?f.then(n):n())}catch(d){return Promise.reject(d)}}(H(),t)}catch(t){return Promise.reject(t)}}({prodUrl:n,host:i,repo:o,floodgatecolor:u,languageFirst:f})).then(function(t){let{caasMetadata:e,errors:r,tags:i,tagErrors:o}=t;const a=e&&!r?.length?function(t,e){void 0===e&&(e=null);let n=null;if(e)try{const t=new URL(e);n=P(t.hostname,t.pathname)}catch(t){n=P()}else n=P();return{entityId:t.entityid,contentId:t.contentid,contentType:t.contenttype,environment:t.env,url:t.url,floodGateColor:t.floodgatecolor,universalContentIdentifier:t.uci,title:t.cardtitle||t.title,description:t.carddescription||t.description,createdDate:t.created,modifiedDate:t.modified,tags:t.tags,primaryTag:t.primarytag,...t.cardimage&&{thumbnail:{altText:t.cardimagealttext,url:t.cardimage}},country:t.country,language:t.lang,cardData:{style:t.style,headline:t.cardtitle||t.title,...t.details&&{details:t.details},...(t.bookmarkenabled||t.bookmarkicon||t.bookmarkaction)&&{bookmark:{enabled:t.bookmarkenabled,bookmarkIcon:t.bookmarkicon,action:t.bookmarkaction}},badges:F(t),...t.playurl&&{playUrl:t.playurl},...(t.cta1url||t.cta2url)&&{cta:{...t.cta1url&&{primaryCta:{text:t.cta1text,url:t.cta1url,style:t.cta1style,icon:t.cta1icon,target:t.cta1target}},...(t.cta2url||"button"===t.cta2style&&t.cta2text)&&{secondaryCta:{text:t.cta2text,url:"button"===t.cta2style?t.cta1url:t.cta2url,style:t.cta2style,icon:t.cta2icon,target:t.cta2target}}}},...(t.eventduration||t.eventstart||t.eventend)&&{event:{duration:t.eventduration,startDate:t.eventstart,endDate:t.eventend}}},origin:t.origin,...t.arbitrary?.length&&{arbitrary:t.arbitrary},...n&&{gbExperienceID:n}}}(e,n):null;return{caasProps:a,caasMetadata:e,errors:r,tags:i,tagErrors:o}})})}catch(t){return Promise.reject(t)}},t.hasCardMetadata=t=>!!t?.querySelector?.(".card-metadata"),t.isDisabledOnPage=t=>{if(!t?.querySelector)return!1;const e=t.querySelector(".card-metadata");return!!e&&[...e.querySelectorAll(":scope > div")].some(t=>{const e=t.querySelectorAll(":scope > div");if(e.length<2)return!1;const n=e[0].textContent.trim().toLowerCase(),r=e[1].textContent.trim().toLowerCase();return"auto-publish"===n&&"false"===r})}});
